requirements.txt
absl-py==2.4.0
accelerate==1.13.0
antlr4-python3-runtime==4.9.3
audioread==3.0.1
beautifulsoup4==4.14.3
bitarray==3.8.0
bitstring==4.4.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
chatterbox-tts @ git+https://github.com/rsxdalv/chatterbox.git@db8b5b75649a51de5405fdee25794d9f58cc09c8
click==8.3.1
conformer==0.3.2
contourpy==1.3.3
cryptography==46.0.7
cssselect==1.4.0
cycler==0.12.1
decorator==5.2.1
diffusers==0.35.1
dill==0.4.0
distlib==0.4.0
einops==0.8.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
fonttools==4.62.1
fsspec==2025.7.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
greenlet==3.3.2
grpcio==1.78.0
hf-xet==1.1.9
httplib2==0.31.2
huggingface-hub==0.34.4
identify==2.6.13
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==8.7.0
jieba3k==0.35.1
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.5.0
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==6.0.2
lxml-html-clean==0.4.4
markdown==3.10.2
markupsafe==3.0.2
matplotlib==3.10.8
ml-dtypes==0.5.3
moviepy==2.2.1
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.1
multiprocess==0.70.18
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.4
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.19.0
packaging==25.0
pandas==3.0.1
peft==0.18.1
pillow==11.3.0
platformdirs==4.4.0
playwright==1.58.0
pooch==1.8.2
praat-parselmouth==0.4.7
pre-commit==4.3.0
proglog==0.1.12
proto-plus==1.27.2
protobuf==6.32.0
psutil==7.2.2
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==2.22
pydub==0.25.1
pyee==13.0.1
pygments==2.19.2
pyloudnorm==0.2.0
pyparsing==3.3.2
pyrubberband==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywavelets==1.9.0
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
resampy==0.4.3
resemble-perth @ git+https://github.com/notdaniel/Perth@8e0cf31f406690ffc256b9fd732b8cc4be40a7a6
s3tokenizer==0.2.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
semchunk==3.2.3
setuptools==82.0.1
sgmllib3k==1.0.0
six==1.17.0
soundfile==0.13.1
soupsieve==2.8.3
sox==1.5.0
soxr==0.5.0.post1
sympy==1.14.0
tabulate==0.10.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tibs==0.5.7
tiktoken==0.11.0
tinysegmenter==0.3
tldextract==5.3.1
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.55.4
triton==3.4.0
typing-extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.34.0
werkzeug==3.1.7
zipp==3.23.0
